Adjust shared packages for web and mobile versions
