+++ FastFoodOrdering/packages/data/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@fastfoodordering/data",
+  "version": "1.0.0",
+  "main": "src/index.ts"
+}
