+++ FastFoodOrdering/packages/data/src/index.ts
@@ -1,0 +1,2 @@
+// This line imports your hook from its file and immediately exports it
+export * from './mockData';
