+++ FastFoodOrdering/packages/data/src/mockData.ts
