+++ FastFoodOrdering/packages/hooks/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@fastfoodordering/hooks",
+  "version": "1.0.0",
+  "main": "src/index.ts"
+}
