+++ FastFoodOrdering/packages/hooks/src/index.ts
@@ -1,0 +1,3 @@
+// This line imports your hook from its file and immediately exports it
+export * from './usePopularItems';
+export * from './useAppState';
