+++ FastFoodOrdering/packages/types/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@fastfoodordering/types",
+  "version": "1.0.0",
+  "main": "src/index.ts"
+}
